pyproject.toml
[project]
name = "hantavirus-ingest"
version = "0.1.0"
description = "Ingest pipeline for the hantavirus tracker — normalizes CDC, WHO, ECDC, PAHO, ProMED, HealthMap into GeoJSON."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "python-dateutil>=2.9",
    "pydantic>=2.10",
    "typer>=0.15",
    "rich>=13.9",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
